Fix Logging not compiled

+++ components/DeviceModule/src/WindowDevice.cpp
@@ -229,7 +229,7 @@ uint16_t WindowDevice::getAttributeUint16Value(uint32_t attributeId) const
     esp_matter_attr_val_t attrVal;
     if (esp_matter::attribute::get_val(attribute, &attrVal) != ESP_OK)
     {
-        ESP_LOGE(TAG, "Failed to get attribute value for ID %d", attributeId);
+        ESP_LOGE(TAG, "Failed to get attribute value for ID %d", (int) attributeId);
         return 0;
     }
